video-sum-mcp/requirements.txt
click>=8.0.0
bilibili-api-python>=15.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
jieba>=0.42.1
python-dotenv>=0.21.0
scikit-learn>=1.2.0
networkx>=2.8.0
numpy>=1.20.0
rich>=13.0.0
matplotlib>=3.5.0

# 异步HTTP客户端依赖
aiohttp>=3.9.0
colorama>=0.4.6

# MCP (Model Context Protocol) 相关依赖
mcp>=0.6.0

# 可选OCR依赖 (安装时可选)
paddlepaddle>=3.0.0
paddleocr>=3.0.0
pillow>=8.0.0

# 可选ASR依赖 (安装时可选)
# openai-whisper>=20230918

# 注意：建议使用Anaconda创建Python环境
# conda create -n py310 python=3.10
# conda activate py310
# pip install -r requirements.txt